requirements-dev.txt
django~=2.2.10
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
pillow
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.3